tests edited and added IUnspentTransactionOutput

+++ src/NovaKey.ExternalApi/Novakey.ExternalApi.BlockChainApi/Models/BlockCypherModels/UnspentTransactionOutputModels/UnspentTransactionOutput.cs
@@ -22,7 +22,7 @@ namespace NovaKey.ExternalApi.BlockChainApi.Models.BlockCypherModels.UnspentTran
         [JsonPropertyName("confirmations")]
         public uint Confirmations { get; set; }
 
-        public decimal GetAmount()
+        public long GetAmount()
         {
             return Amount;
         }
